Remove is-image-url dependency
Implemented isImageUrl utility func which
is much simpler and doesn't bring so many dependencies.

Use `useLayoutEffect` rather than `useEffect` for setting
the image src from image url, mainly because (currently)
`useEffect` can't be tested in our case. The thing is that
effects are executed after the render is committed to the screen,
while layout effect are run before the DOM is painted.